+++ tests/test_input_service.py
@@ -70,18 +70,3 @@ def test_is_not_done_adding_with_items():
 
 def test_is_not_done_adding_with_existing_item():
     assert is_done_adding_items("foobar", [("foobar", 1)]) == False
-
-def test_parse_silo_count():
-    assert parse_silo_count("1") == 1
-    assert parse_silo_count("5") == 5
-    assert parse_silo_count("20") == 20
-
-def test_parse_silo_count_non_int():
-    with pytest.raises(ValueError, match=INPUT_INVALID_NUM):
-        parse_silo_count("foobar")
-
-def test_parse_silo_count_invalid_int():
-    with pytest.raises(ValueError, match=INPUT_GREATER_ZERO):
-        parse_silo_count("0")
-    with pytest.raises(ValueError, match=INPUT_GREATER_ZERO):
-        parse_silo_count("-1")
